added option to allow all elements
Added that when the checkbox is pressed, the options are updated too, as well as a text that appears to warn about potential issues when checking this option. Tried to do a warning but decided against it because of some quirks and complications. The errors that houdini throws when it can't read a file is very detailed and self explanatory already.